+++ app/src/main/res/layout/fragment_profile.xml
@@ -1,14 +1,102 @@
 <?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
+
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ProfileFragment">
+    android:orientation="vertical"
+    tools:context=".view.ProfileFragment">
 
-    <!-- TODO: Update blank fragment layout -->
-    <TextView
+    <LinearLayout
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="@string/hello_blank_fragment" />
+        android:layout_height="wrap_content"
+        android:gravity="center"
+        android:orientation="horizontal"
+        android:paddingHorizontal="24dp"
+        android:paddingVertical="10dp">
 
-</FrameLayout>
+
+        <com.google.android.material.card.MaterialCardView
+
+            android:layout_width="100dp"
+            android:layout_height="100dp"
+
+            app:cardCornerRadius="50dp"
+            app:cardElevation="0dp">
+
+            <ImageView
+                android:id="@+id/register_activity_profile_picture"
+                android:layout_width="match_parent"
+                android:layout_height="match_parent"
+                android:layout_margin="15dp"
+                android:contentDescription=""
+                android:scaleType="fitXY"
+                />
+
+        </com.google.android.material.card.MaterialCardView>
+
+        <Space
+            android:layout_width="40dp"
+            android:layout_height="match_parent" />
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical">
+
+            <TextView
+                android:id="@+id/fragment_profile_name"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="sample name"
+
+                tools:ignore="hardCodedText" />
+
+            <TextView
+                android:id="@+id/fragment_profile_nickname"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:layout_marginBottom="5dp"
+                android:text="sample nick"
+
+                tools:ignore="hardCodedText" />
+
+            <TextView
+                android:id="@+id/fragment_profile_pet_preference"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="sample pet"
+
+                tools:ignore="hardCodedText" />
+
+            <TextView
+                android:id="@+id/fragment_profile_job_status"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="sample job"
+
+                tools:ignore="hardCodedText" />
+
+            <com.google.android.material.button.MaterialButton
+                style="@style/Widget.MaterialComponents.Button.TextButton"
+                android:layout_width="match_parent"
+                android:layout_height="wrap_content"
+                android:text="Update"
+                android:visibility="gone" />
+
+        </LinearLayout>
+
+    </LinearLayout>
+
+    <com.google.android.material.divider.MaterialDivider
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_marginBottom="10dp" />
+
+    <androidx.recyclerview.widget.RecyclerView
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:paddingHorizontal="24dp" />
+
+</LinearLayout>
